Enable edits to the table to be saved when clicking "OK"
This makes it so it isn't necessary for the user to click the "Edit"
button for each API provider the user wants to edit. Instead, they can
simply edit the cells in the table directly and click "OK" to save all
their changes at once. Also, after editing a cell, if the user needs a
larger text field to type in they can now click the "Edit" button and
the changes they made in the table will be reflected in the API Provider
dialogue.

This works by creating a table model listener that listens for update
events, and updates the corresponding provider whenever a cell in the
table is updated. We also create a custom table model that trims the
whitespace in string fields whenever the setValueAt method is called, so
we can guarantee that any time the value of a String cell is set, the
cell won't accidentally contain a bunch of extra whitespace. This works
for both direct cell edits and setting the value from code.