Add syntax highlighting to code blocks
This code adds a custom NodeRenderer extension that uses JRuby to run
Rouge to perform syntax highlighting on fenced code blocks. The syntax
highlighting makes it easier to understand large blocks of code at a
glance.

Unfortunately this has major performance issues. The first major issue
is that it takes several seconds for JRuby to start up, so the first
time the "Explain Function" (etc.) button is clicked, Ghidra will freeze
for those seconds while JRuby initializes. This is annoying, but
relatively low-impact since that start-up cost is only paid once per
Ghidra session, and I assume most users aren't repeatedly closing and
re-opening Ghidra.

The second issue is much more impactful. When processing longer
responses from the LLM, Ghidra starts freezing as it gets to the end of
the message. I think this is happening because it's re-rendering the
syntax highlighting for every new chunk of text. So as the full message
gets longer it takes correspondingly longer to re-render it, and
eventually it takes so long to render that it can't be finished before
the next chunk arrives. At that point, it just gets completely stuck
until the arrival of the final chunk.

Maybe there's a way to make it so the chunks can still be received and
kept in a buffer while the full message is rendering? So when the
rendering finishes it can append all the new chunks to the message
before it tries to render it, rather than appending and re-rendering
each chunk individually.